+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+## Added
+- Exception to step result ([#78](https://github.com/cucumber/fake-cucumber/pull/78))
+- Upgrade to `@cucumber/messages` `21`
+- Upgrade to `@cucumber/gherkin` `26`
 
 ## [16.0.0] - 2022-06-01
 ### Added